paddleocr-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
paddleocr>=2.7.3
paddlepaddle>=3.3.0
pillow>=10.4.0
pydantic>=2.5.0
python-multipart>=0.0.6
